test-web: use path with unique unicode sequence
Change the test path from "éxcitement" to "¡excitement!" since the
former has multiple valid unicode encodings (e.g. given combining
characters, etc.).

With the previous string, the test might fail if something along the way
normalized it (which was the case on at least some versions of OS X).

Thanks to ilovezfs for reporting the problem and helping with the fix.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>