test-gc: sort ls output defensively before comm
Explicitly sort the output of ls before passing it to comm in order to
accommodate some systems that apparently used to change the default sort
order.  The order must match LC_COLLATE for comm to work.

While we're here, improve error checking around the comm invocations.

Thanks to Nick Alcock for reporting the original problem, and proposing
an initial fix.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>