vfs2: suffix duplicate save names (e.g. 1970-01-01-214640-07)
From now on, append zero-padded integers to the names of saves that
have the same commit time in order to avoid duplicates.  The sequence
number currently represents the save's reversed position in default
rev-list order, so that given:

  /foo/1970-01-01-214640-09
  /foo/1970-01-01-214640-10

In the normal case, the -10 save would indicate the next save made
after -09 (and the -09 save would be the single parent commit for
-10).

Thanks to Ben Kelley for helping me sort out some of the details.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>