restore: create fifos with mkfifo, not mknod
I recently did a restore of a large bup backup, about 34G worth.  All
worked well, including metadata, except that bup threw an exception on
restoring fifos (that I didn't need; they were in /var and were sockets
in use by daemons when the backup happened).

The problem was that mknod was being called for the fifo, and given only
two argumetns.  mknod(2) on NetBSD says it takes three arguments.
mkfifo takes two.  I am guessing that mknod in python calls mknod the OS
call, and on Linux somehow the third null argument works out somehow.
But it seems irregular to make a fifo with mknod.

I realize python is not POSIX, but mknod(2) requires three arguments:
  http://pubs.opengroup.org/onlinepubs/9699919799/functions/mknod.html

It would be nice to have a test of backing up and restoring a fifo; that
would have caught this.

The following patch makes my restore go smoothly.

Signed-off-by: Greg Troxel <gdt@lexort.com>
[rlb@defaultvalue.org: adjust commit summary]
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>