Move bup to cmd/ and symlink ./bup to cmd/bup
Move main.py to cmd/bup and make ./bup a permanent symlink to it.
Make the installed bup executable a relative symlink to the
LIBDIR/cmd/bup file.  Always use the location of the bup executable as
a way to locate the other subcommands and the lib and resource
directories, either when running from the source tree or from an
install tree.

This work finishes the switch to have all the commands to rely on the
bup-python wrapper so that we can establish some norms we'll need to
support Python 3.  It should also allow us to simplify the main.py
startup process.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>