Install bup-python and rely on it everywhere
Add a cmd/python-cmd.sh template and use it to generate
cmd/bup-python.  Install bup-python alongside all the other commands
and stop rewriting the command #! headers at install time, so that
they'll use it.

We're setting up all of the commands to rely on bup-python so that
we'll have a common place to establish some norms that we'll want in
order to support Python 3.

Though in fact, you can't actually run the python-based subcommands
independently because they don't have the right PYTHONPATH, etc.
That's something we'll also fix.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>