server: change protocol/backend relationship to composition
For the conversion, it made sense to have the concrete backend
inherit from the protocol server, but now that it's split it
actually makes more sense to just compose them, so we can also
swap out the protocol for example.

Make that change, but keep an AbstractServerBackend class that
serves mostly to hold the documentation.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>