server: improve repo handling on multiple commands
Currently, we re-instantiate the repo object on every single
command, including when a receive-pack-v2 command was suspended.
This is not only inefficient, it also precludes moving pack
writing state into the repo object, since that would then break
with the suspended write.

Additionally, there's a bug in that you can theoretically
suspend a receive-pack-v2 command, switch the repository using
set-dir and then resume the receive-pack-v2, still writing to
the old repository.

Fix that as well, by closing the suspended_w when switching
the repo object to a new one.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>