bupsplit: implement RecordHashSplitter class
Instances of this class can be used to apply hashsplitting to
a stream of records, .feed() the object the records one by one,
and it'll return if splitting should be done. If yes, it resets
the internal state to start fresh from the next record.

Prior to this, the various rollsum_* functions were only used
in the bupsplit.c file, which meant that they got inlined all
the time, which is good for performance here. Since I now use
them in _helpers.c as well, declare them inlines in the header
file directly.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Signed-off-by: Rob Browning <rlb@defaultvalue.org>

[rlb@defaultvalue.org: adjusted some types to match current
 main (size_t, etc.); adjusted to use bits_from_py_kw();
 return a None offset from feed() when no split.]

Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>

---
v2:
 * rebase to not require all my other patches
v3:
 * correctly rebase, removing parts of the code that were
   referencing hashsplit with permutation
v4:
 * inline everything to recover speed
v5:
 * combine with HashSplitter object, etc.