bup: use correct bup executable in on--server
Currently, on--server loses the correct bup command in one situation:
if you run 'bup on remote' with a remote forced ssh command that's
different from the installed bup, on--server will still execute the
installed version of bup for all sub-commands.

Fix that by using bup.path.exe() in on--server for bup execution.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
---
v2:
 * use bup.path.exe()
v3:
 * rebase on py3 changes