git: PackIdxList/PackIdx: add the ability to return the offset
Add the ability to not just query _whether_ an object exists, and
if so return in which pack (or midx), but also to return the offset
at which it exists.

Note that this is (currently) incompatible with midx searches since
another search in the corresponding idx would be required and that
makes the code more complex, hopefully the bloom filter helps enough,
for now just skip any midx if want_offs is set.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>