git: use git's int parsing with pack.packSizeLimit config
Our parse_num() understands a little more than git, in particular
 * T for terabytes
 * b suffix when you specify Kb, Mb, Gb or Tb.

Neither of those are understood by git, it only understands the
K, M and G suffixes (case-insensitive). However, a git repository
that actually states 'pack.packSizeLimit' as something that git
doesn't understand is broken for every single git command, and as
we call git, the added flexibility of parse_num() cannot be used.

Thus, simplify the code and just use opttype='int' for this.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>