Add close_fds to readpipe batchpipe, and exo
Add close_fds to readpipe for import-duplicity, and while we're
there, define readpipe in terms of exo(), since it's just a
simplification.  Add close_fds to batchpipe since batchpipe is
intended to just be a readpipe for long argument lists.

Default close_fds to True, since it doesn't look like we need to
preserve the open fds in our other calls.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>