metadata: accept only fixed python-xattr in python3
This is currently broken on python3, it returns junk when
we pass bytes because it uses string and %s internally.
I made a fix for it, so we can detect here if it's fixed
(in which case the NS_USER constant is bytes, not string),
load the module only if it is indeed fixed.

We can do this test always since in python2 bytes == str
and thus isinstance('user', bytes) == True.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
---
v2:
 * accept python-xattr if pyxattr_compat.NS_USER is bytes

v3:
 * fix commit message

v4:
 * remove pointless version check