web: generalize query parameter handling (e.g. "hidden")
In preparation for additional parameters, generalize the handling of
the existing "hidden" parameter.  Add ParamInfo to keep track of the
relevant conversions and defaults.  Use it to parse the incoming
request values and to render url query components.

Begin converting the code to encode url path parts via urllib quote
and parameters via urllib urlencode (respecting the "+ difference for
spaces"), and to encode the parts on the page via path_msg.

Adapted from previous work by Johannes Berg.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>